requirements.txt
fastapi
uvicorn[standard]
requests

sqlalchemy
psycopg2-binary

chromadb
sentence-transformers

gitpython

groq

redis
python-dotenv

slowapi
limits

pydantic

passlib[bcrypt]
bcrypt

pytest
alembic

black
ruff
mypy